Conflicts:
	WalMart/WalMart.xcodeproj/project.pbxproj
	WalMart/WalMart/Images.xcassets/LaunchImage.launchimage/Contents.json
	WalMart/WalMart/Shared/Services/ShoppingCart/GRShoppingCartProductsService.swift
	WalMart/WalMart/iPad/View/SearchProduct/IPASearchProductCollectionViewCell.swift
	WalMart/WalMart/iPhone/Controller/CheckOut/GRCheckOutViewController.swift
	WalMart/WalMart/iPhone/Controller/Navigation/MoreOptionsViewController.swift
	WalMart/WalMart/iPhone/View/SearhResult/SectionHeaderSearchHeader.swift

+++ WalMart/WalMart/Shared/Services/LoginSignUp/LoginWithEmailService.swift
@@ -13,42 +13,52 @@ class LoginWithEmailService : BaseService {
     
     var loginIdGR = ""
     
+    
     func buildParams(email:String,password: String) -> NSDictionary {
         let lowCaseUser = email.lowercaseString
         return ["email":lowCaseUser]
     }
     
     func callService(params:NSDictionary,successBlock:((NSDictionary) -> Void)?, errorBlock:((NSError) -> Void)? ) {
-        self.callPOSTService(params, successBlock: { (resultCall:NSDictionary) -> Void in
-            if let codeMessage = resultCall["codeMessage"] as? NSNumber {
-                if codeMessage.integerValue == 0 &&  UserCurrentSession.hasLoggedUser(){
-                    let resultCallMG = resultCall
-                    let cadUserId : NSString? = UserCurrentSession.sharedInstance().userSigned!.idUserGR
-                    if cadUserId != nil && cadUserId != "" && cadUserId?.length > 0 {
-                        let serviceGr = GRLoginService()
-                        serviceGr.callService(serviceGr.buildParamsUserId(), successBlock:{ (resultCall:NSDictionary?) in
-                            UserCurrentSession.sharedInstance().createUpdateUser(resultCallMG, userDictionaryGR: resultCall!)
-                            successBlock!(resultCall!)
-                            }
-                            , errorBlock: {(error: NSError) in
-                                errorBlock!(error)
-                        })
-                    }else {
-                        UserCurrentSession.sharedInstance().userSigned = nil
-                        UserCurrentSession.sharedInstance().deleteAllUsers()
+        if !UserCurrentSession.sharedInstance().userSignedOnService {
+            UserCurrentSession.sharedInstance().userSignedOnService = true
+            self.callPOSTService(params, successBlock: { (resultCall:NSDictionary) -> Void in
+                if let codeMessage = resultCall["codeMessage"] as? NSNumber {
+                    if codeMessage.integerValue == 0 &&  UserCurrentSession.hasLoggedUser(){
+                        let resultCallMG = resultCall
+                        let cadUserId : NSString? = UserCurrentSession.sharedInstance().userSigned!.idUserGR
+                        if cadUserId != nil && cadUserId != "" && cadUserId?.length > 0 {
+                            let serviceGr = GRLoginService()
+                            serviceGr.callService(serviceGr.buildParamsUserId(), successBlock:{ (resultCall:NSDictionary?) in
+                                UserCurrentSession.sharedInstance().createUpdateUser(resultCallMG, userDictionaryGR: resultCall!)
+                                successBlock!(resultCall!)
+                                UserCurrentSession.sharedInstance().userSignedOnService = false
+                                }
+                                , errorBlock: {(error: NSError) in
+                                    errorBlock!(error)
+                                    UserCurrentSession.sharedInstance().userSignedOnService = false
+                            })
+                        }else {
+                            UserCurrentSession.sharedInstance().userSigned = nil
+                            UserCurrentSession.sharedInstance().deleteAllUsers()
+                        }
+                    }
+                    else{
+                        let errorDom = NSError(domain: "com.bcg.service.error", code: 0, userInfo: nil)
+                        //let message = resultCall["message"] as! String
+                        //let error = NSError()
+                        //error.setValue(message, forKey:codeMessage)
+                        errorBlock!(errorDom)
                     }
                 }
-                else{
-                    let errorDom = NSError(domain: "com.bcg.service.error", code: 0, userInfo: nil)
-                    //let message = resultCall["message"] as! String
-                    //let error = NSError()
-                    //error.setValue(message, forKey:codeMessage)
-                    errorBlock!(errorDom)
-                }
+                }) { (error:NSError) -> Void in
+                    errorBlock!(error)
+                    UserCurrentSession.sharedInstance().userSignedOnService = false
             }
-            }) { (error:NSError) -> Void in
-                errorBlock!(error)
+        } else {
+            successBlock?([:])
         }
+        
     }
     
     override func shouldIncludeHeaders() -> Bool {
